requirements.txt
# AI Learning Assistant Backend Requirements
# Install with: pip install -r requirements.txt

# Web framework
flask==2.3.3
flask-cors==4.0.0

# Document processing
PyPDF2==3.0.1
python-pptx==0.6.21

# AI and ML libraries
openai==0.28.1
langchain==0.0.304

# Vector storage and embeddings
faiss-cpu==1.7.4
tiktoken==0.5.1

# Additional utilities
python-dotenv==1.0.0
requests==2.31.0

# Production deployment (optional)
gunicorn==21.2.0